backend/requirements.txt
# ============================================
# Retail Media Creative Builder - Requirements
# Hybrid Mode: Works locally AND in cloud (Render/Railway/Vercel)
# ============================================

# ===================
# Web Framework (REQUIRED)
# ===================
Flask>=3.0.0
Flask-CORS>=4.0.0
Werkzeug>=3.0.0

# ===================
# AI/LLM Chat (REQUIRED)
# ===================
google-genai                # Google Gemini API for AI chat

# ===================
# Image Generation - HYBRID MODE
# ===================
# Uses Hugging Face API by default (free, no downloads, cloud-friendly)
# Optional: Install torch+diffusers for local generation (10GB+ downloads)
# 
# For LOCAL mode (unlimited generation):
#   pip install torch torchvision diffusers transformers accelerate
#
# For API mode (cloud deployment):
#   Already configured - uses free Hugging Face API

# ===================
# Background Removal (REQUIRED)
# ===================
rembg>=2.0.0                # AI background removal
onnxruntime>=1.16.0         # ONNX model inference

# ===================
# Image Processing (REQUIRED)
# ===================
Pillow>=10.0.0              # Image manipulation
numpy>=1.24.0,<2.0.0        # Numerical operations

# ===================
# HTTP & Async (REQUIRED)
# ===================
requests>=2.31.0            # HTTP requests for API calls
httpx>=0.25.0               # Async HTTP client
aiohttp>=3.9.0              # Async HTTP
aiofiles>=23.0.0            # Async file operations

# ===================
# Data Validation (REQUIRED)
# ===================
pydantic>=2.5.0             # Data validation
pydantic-settings>=2.1.0    # Settings management

# ===================
# Utilities (REQUIRED)
# ===================
loguru>=0.7.0               # Better logging
python-dotenv>=1.0.0        # Environment variables (.env file support)
python-multipart>=0.0.6     # File uploads
tqdm>=4.66.0                # Progress bars
colorama>=0.4.6             # Colored terminal output
PyYAML>=6.0.0               # YAML config files
piexif>=1.1.3               # Image metadata
filetype>=1.2.0             # File type detection

# ===================
# Optional - Local Image Generation (10GB+ downloads)
# ===================
# Uncomment these for local unlimited image generation:
# torch>=2.0.0
# torchvision>=0.15.0
# diffusers>=0.25.0
# transformers>=4.35.0
# accelerate>=0.25.0
# safetensors>=0.4.0
# huggingface_hub>=0.19.0
# tokenizers>=0.15.0
# sentencepiece>=0.1.99

# ===================
# Optional - Advanced Features
# ===================
# opencv-python>=4.8.0      # Advanced image processing (heavy)
# scikit-image>=0.22.0      # Image algorithms (heavy)
# scipy>=1.11.0             # Scientific computing
# matplotlib>=3.8.0         # Plotting/debugging

